+++ app/models/sequel_form_compatibility.rb
@@ -1,5 +1,0 @@
-module SequelFormCompatibility
-  def model_name
-    self.class.model_name
-  end
-end