+++ NEWS.md
@@ -63,6 +63,7 @@ Development
 * Revamp Rubocop config
 * Fix visualization backup when permission is missing [#15874](https://github.com/CartoDB/cartodb/pull/15874)
 * Show outdated subscriptions. Optimize requests [#15879](https://github.com/CartoDB/cartodb/pull/15879)
+* Include `::Carto::ActiveRecordCompatibility` in all `Sequel` models [#15879](https://github.com/CartoDB/cartodb/pull/15879)
 
 4.41.1 (2020-09-03)
 -------------------